begin refactor to etch
