+++ lib/views/brand.js
@@ -1,10 +1,47 @@
+'use babel';
+/** @jsx etch.dom */
+const etch = require('etch');
+
 module.exports = {
-  brightGold: 'E7BC70',
-  // coffeelint: disable=max_line_length
-  tuxFullColor: "<svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 89 89.6'><style>.a{fill:#D2AB67;}.b{fill:#E7BC70;}.c{fill:#060500;}.d{fill:#2F2D29;}.e{fill:#7E868C;}</style><polygon points='73.4 37.6 88.6 50.7 44.4 89.2 44.4 89.1 44.4 89.2 0.4 50.9 15.7 37.7 44.5 79.9 ' class='a'/><polygon points='44.5 89.1 88.6 50.7 73.4 37.6 44.5 79.9 ' class='b'/><polygon points='81.9 21.6 44.5 76.5 7.1 21.6 44.5 0.4 ' class='a'/><path d='M44.5 76.4l37.3-54.8 -6 8.6H44.5V76.4zM56.5 52.9l-0.3-19.6 19.4-2.8L56.5 52.9z' class='b'/><polygon points='44.5 0.4 44.5 10.7 47.4 12.1 47.4 15 44.5 16.5 44.5 21.6 81.9 21.6 ' class='b'/><polygon points='81.9 21.6 44.5 67.7 7.1 21.6 44.5 4.8 ' class='c'/><polygon points='44.5 4.8 44.5 4.8 44.5 67.6 44.5 67.7 81.9 21.6 ' class='d'/><polygon points='13.1 30.3 75.8 30.3 81.9 21.6 7.1 21.6 ' class='a'/><polygon points='44.5 21.6 44.5 30.3 75.8 30.3 81.9 21.6 ' class='b'/><polygon points='65.6 25.9 23.4 25.9 20.4 21.5 68.5 21.5 ' class='c'/><polygon points='44.5 21.5 44.5 25.9 65.6 25.9 68.6 21.5 ' class='d'/><polygon points='22.5 24.4 23.4 25.9 65.6 25.9 66.5 24.4 ' class='e'/><circle cx='44.5' cy='33.7' r='1.2' class='e'/><circle cx='44.5' cy='38.8' r='1.2' class='e'/><circle cx='44.5' cy='43.8' r='1.2' class='e'/><polygon points='32.5 52.7 32.7 33.2 13.3 30.3 ' class='a'/><polygon points='32.5 52.7 27.5 38.2 13.3 30.3 ' class='c'/><polygon points='56.5 52.9 56.2 33.2 75.6 30.4 ' class='a'/><polygon points='56.2 33.2 56.5 52.9 75.6 30.4 ' class='b'/><polygon points='56.5 52.9 61.6 38.2 75.6 30.4 ' class='d'/><polygon points='47.4 15 44.5 16.5 41.5 15 41.5 12.1 44.5 10.7 47.4 12.1 ' class='a'/><polygon points='44.5 10.7 44.5 16.5 47.4 15 47.4 12.1 ' class='b'/><path d='M45.2 70.8c0 0.3-0.3 0.6-0.6 0.6h-0.3c-0.3 0-0.6-0.3-0.6-0.6v-22.3c0-0.3 0.3-0.6 0.6-0.6h0.3c0.3 0 0.6 0.3 0.6 0.6V70.8z' class='e'/></svg>",
+    brightGold: 'E7BC70',
+    // coffeelint: disable=max_line_length
+    tuxFullColor: (
+        <svg viewBox="0 0 89 89.6">
+            <polygon className="a" points="73.4 37.6 88.6 50.7 44.4 89.2 44.4 89.1 44.4 89.2 0.4 50.9 15.7 37.7 44.5 79.9 "/>
+            <polygon className="b" points="44.5 89.1 88.6 50.7 73.4 37.6 44.5 79.9 "/>
+            <polygon className="a" points="81.9 21.6 44.5 76.5 7.1 21.6 44.5 0.4 "/>
+            <path className="b" d="M44.5 76.4l37.3-54.8 -6 8.6H44.5V76.4zM56.5 52.9l-0.3-19.6 19.4-2.8L56.5 52.9z"/>
+            <polygon className="b" points="44.5 0.4 44.5 10.7 47.4 12.1 47.4 15 44.5 16.5 44.5 21.6 81.9 21.6 "/>
+            <polygon className="c" points="81.9 21.6 44.5 67.7 7.1 21.6 44.5 4.8 "/>
+            <polygon className="d" points="44.5 4.8 44.5 4.8 44.5 67.6 44.5 67.7 81.9 21.6 "/>
+            <polygon className="a" points="13.1 30.3 75.8 30.3 81.9 21.6 7.1 21.6 "/>
+            <polygon className="b" points="44.5 21.6 44.5 30.3 75.8 30.3 81.9 21.6 "/>
+            <polygon className="c" points="65.6 25.9 23.4 25.9 20.4 21.5 68.5 21.5 "/>
+            <polygon className="d" points="44.5 21.5 44.5 25.9 65.6 25.9 68.6 21.5 "/>
+            <polygon className="e" points="22.5 24.4 23.4 25.9 65.6 25.9 66.5 24.4 "/>
+            <circle className="e" cx="44.5" cy="33.7" r="1.2"/>
+            <circle className="e" cx="44.5" cy="38.8" r="1.2"/>
+            <circle className="e" cx="44.5" cy="43.8" r="1.2"/>
+            <polygon className="a" points="32.5 52.7 32.7 33.2 13.3 30.3 "/>
+            <polygon className="c" points="32.5 52.7 27.5 38.2 13.3 30.3 "/>
+            <polygon className="a" points="56.5 52.9 56.2 33.2 75.6 30.4 "/>
+            <polygon className="b" points="56.2 33.2 56.5 52.9 75.6 30.4 "/>
+            <polygon className="d" points="56.5 52.9 61.6 38.2 75.6 30.4 "/>
+            <polygon className="a" points="47.4 15 44.5 16.5 41.5 15 41.5 12.1 44.5 10.7 47.4 12.1 "/>
+            <polygon className="b" points="44.5 10.7 44.5 16.5 47.4 15 47.4 12.1 "/>
+            <path className="e" d="M45.2 70.8c0 0.3-0.3 0.6-0.6 0.6h-0.3c-0.3 0-0.6-0.3-0.6-0.6v-22.3c0-0.3 0.3-0.6 0.6-0.6h0.3c0.3 0 0.6 0.3 0.6 0.6V70.8z"/>
+        </svg>
+    ),
 
-  tuxSingleColor: "<svg xmlns='http://www.w3.org/2000/svg' width='89' height='90' viewBox='0 0 89 89.6'><polygon points='73.4 35.4 88.6 48.5 44.4 87 44.4 86.9 44.4 87 0.4 48.7 15.7 35.5 44.5 77.7 ' class='a'/><polygon points='22.5 22.2 23.4 23.7 65.6 23.7 66.5 22.2 ' class='a'/><polygon points='32.5 50.5 27.5 36 13.3 28.1 ' class='a'/><polygon points='56.5 50.7 61.6 36 75.6 28.2 ' class='a'/><path d='M44.5 2.6L7.1 19.4h13.4l2.8 4.3h42.2l2.8-4.3h13.4L44.5 2.6zM47.4 12.9l-2.9 1.5 -2.9-1.5V9.9l2.9-1.5 2.9 1.5V12.9z' class='a'/><path d='M74.8 28.1H14.1l0.1 0.1 18.5 2.8 -0.3 19.5 -2.4-2.8 13.7 16.8v-18.3c0-0.3 0.3-0.6 0.6-0.6h0.3c0.3 0 0.6 0.3 0.6 0.6v18.3l11.4-14 -0.1 0.1 -0.3-19.6L74.6 28.4 74.8 28.1zM44.5 42.8c-0.6 0-1.2-0.5-1.2-1.2 0-0.6 0.5-1.2 1.2-1.2 0.6 0 1.2 0.5 1.2 1.2C45.6 42.2 45.1 42.8 44.5 42.8zM44.5 37.8c-0.6 0-1.2-0.5-1.2-1.2 0-0.6 0.5-1.2 1.2-1.2 0.6 0 1.2 0.5 1.2 1.2C45.6 37.2 45.1 37.8 44.5 37.8zM44.5 32.6c-0.6 0-1.2-0.5-1.2-1.2s0.5-1.2 1.2-1.2c0.6 0 1.2 0.5 1.2 1.2S45.1 32.6 44.5 32.6z' class='a'/></svg>",
+    tuxSingleColor: '<svg xmlns="http://www.w3.org/2000/svg" width="89" height="90" viewBox="0 0 89 89.6"><polygon points="73.4 35.4 88.6 48.5 44.4 87 44.4 86.9 44.4 87 0.4 48.7 15.7 35.5 44.5 77.7 " class="a"/><polygon points="22.5 22.2 23.4 23.7 65.6 23.7 66.5 22.2 " class="a"/><polygon points="32.5 50.5 27.5 36 13.3 28.1 " class="a"/><polygon points="56.5 50.7 61.6 36 75.6 28.2 " class="a"/><path d="M44.5 2.6L7.1 19.4h13.4l2.8 4.3h42.2l2.8-4.3h13.4L44.5 2.6zM47.4 12.9l-2.9 1.5 -2.9-1.5V9.9l2.9-1.5 2.9 1.5V12.9z" class="a"/><path d="M74.8 28.1H14.1l0.1 0.1 18.5 2.8 -0.3 19.5 -2.4-2.8 13.7 16.8v-18.3c0-0.3 0.3-0.6 0.6-0.6h0.3c0.3 0 0.6 0.3 0.6 0.6v18.3l11.4-14 -0.1 0.1 -0.3-19.6L74.6 28.4 74.8 28.1zM44.5 42.8c-0.6 0-1.2-0.5-1.2-1.2 0-0.6 0.5-1.2 1.2-1.2 0.6 0 1.2 0.5 1.2 1.2C45.6 42.2 45.1 42.8 44.5 42.8zM44.5 37.8c-0.6 0-1.2-0.5-1.2-1.2 0-0.6 0.5-1.2 1.2-1.2 0.6 0 1.2 0.5 1.2 1.2C45.6 37.2 45.1 37.8 44.5 37.8zM44.5 32.6c-0.6 0-1.2-0.5-1.2-1.2s0.5-1.2 1.2-1.2c0.6 0 1.2 0.5 1.2 1.2S45.1 32.6 44.5 32.6z" class="a"/></svg>',
 
-  text: "<svg xmlns='http://www.w3.org/2000/svg' width='160' height='35' viewBox='0 0 160 35'><path d='M29.9 12.6c0 7.8-4.9 12.4-13.4 12.4h-5.9v9.2H1.9V0.9h14.6C25 0.9 29.9 5.2 29.9 12.6zM21.6 12.8c0-3.3-2-5-5.5-5h-5.5v10.2h5.5C19.6 18 21.6 16.2 21.6 12.8z'/><path d='M62.5 34.1l-4.8-9.2h-0.2 -6.3v9.2h-8.7V0.9h14.9c8.8 0 13.8 4.3 13.8 11.7 0 5-2.1 8.7-5.9 10.7l6.9 10.8H62.5zM51.3 18h6.3c3.5 0 5.5-1.8 5.5-5.2 0-3.3-2-5-5.5-5h-6.3V18z'/><path d='M119.1 17.5c0 9.7-7.7 17-18 17 -10.3 0-17.9-7.3-17.9-17 0-9.7 7.7-16.8 18-16.8C111.4 0.7 119.1 7.9 119.1 17.5zM92.1 17.5c0 5.4 4.2 9.6 9.1 9.6 5 0 9-4.2 9-9.6s-4-9.5-9-9.5C96.3 8.1 92.1 12.1 92.1 17.5z'/><path d='M144 7.6c-2.1 0-3.5 0.8-3.5 2.3 0 5.5 17.5 2.4 17.5 14.3 0 6.8-6 10.2-13.4 10.2 -5.5 0-11.3-2-15.3-5.3l3.4-6.8c3.4 2.9 8.6 5 12 5 2.6 0 4.2-0.9 4.2-2.7 0-5.6-17.5-2.2-17.5-14 0-6.2 5.3-10.1 13.3-10.1 4.9 0 9.8 1.5 13.3 3.7l-3.3 6.9C151.2 9.1 146.8 7.6 144 7.6z'/></svg>"
+    text: (
+        <svg height="35" width="160" viewBox="0 0 160 35">
+            <path d="M29.9 12.6c0 7.8-4.9 12.4-13.4 12.4h-5.9v9.2H1.9V0.9h14.6C25 0.9 29.9 5.2 29.9 12.6zM21.6 12.8c0-3.3-2-5-5.5-5h-5.5v10.2h5.5C19.6 18 21.6 16.2 21.6 12.8z"/>
+            <path d="M62.5 34.1l-4.8-9.2h-0.2 -6.3v9.2h-8.7V0.9h14.9c8.8 0 13.8 4.3 13.8 11.7 0 5-2.1 8.7-5.9 10.7l6.9 10.8H62.5zM51.3 18h6.3c3.5 0 5.5-1.8 5.5-5.2 0-3.3-2-5-5.5-5h-6.3V18z"/>
+            <path d="M119.1 17.5c0 9.7-7.7 17-18 17 -10.3 0-17.9-7.3-17.9-17 0-9.7 7.7-16.8 18-16.8C111.4 0.7 119.1 7.9 119.1 17.5zM92.1 17.5c0 5.4 4.2 9.6 9.1 9.6 5 0 9-4.2 9-9.6s-4-9.5-9-9.5C96.3 8.1 92.1 12.1 92.1 17.5z"/>
+            <path d="M144 7.6c-2.1 0-3.5 0.8-3.5 2.3 0 5.5 17.5 2.4 17.5 14.3 0 6.8-6 10.2-13.4 10.2 -5.5 0-11.3-2-15.3-5.3l3.4-6.8c3.4 2.9 8.6 5 12 5 2.6 0 4.2-0.9 4.2-2.7 0-5.6-17.5-2.2-17.5-14 0-6.2 5.3-10.1 13.3-10.1 4.9 0 9.8 1.5 13.3 3.7l-3.3 6.9C151.2 9.1 146.8 7.6 144 7.6z"/>
+        </svg>
+    )
 };
-  // coffeelint: enabled=max_line_length
+// coffeelint: enabled=max_line_length
