+++ styles/welcome.less
@@ -136,3 +136,20 @@ div.pros-welcome > div {
   max-width: 50em;
   min-width: 35em;
 }
+svg {
+    .a {
+        fill: #D2AB67;
+    }
+    .b {
+        fill:#E7BC70;
+    }
+    .c {
+        fill: #060500;
+    }
+    .d {
+        fill: #2F2D29;
+    }
+    .e {
+        fill: #7E868C;
+    }
+}
